added a boolean "internal" which indicates that the data object is an
2002-06-12  Michael Natterer  <mitch@gimp.org>

	* app/core/gimpdata.[ch]: added a boolean "internal" which
	indicates that the data object is an automatically created
	internal object. Changed the code to refuse saving to deleting
	internal objects.

	* app/core/Makefile.am
	* app/core/gimp-gradients.[ch]: new files implementing internal
	gradients (FG -> BG in RGB and HSV and FG -> transparent).

	* app/core/gimp.c: call gimp_gradients_init().

	* app/core/gimpdatafactory.c (gimp_data_factory_data_free): don't
	free internal objects so they stay there on "Refresh".

	* app/core/gimpdatalist.c: sort internal objects to the beginning
	of the list.

	* app/widgets/gimpdataeditor.c: refuse to change the name of
	internal objects.

	* app/widgets/gimpdatafactoryview.c: set the "Delete" button
	insensitive for internal objects.

	* app/widgets/gimpgradienteditor.c: refuse to edit internal
	gradients, just display them so color picking works.

	* app/gui/brushes-commands.c
	* app/gui/gradients-commands.c
	* app/gui/palettes-commands.c
	* app/gui/patterns-commands.c: set the "Delete" menu item
	insensitive for internal objects.

	* app/gui/gui.c: need to call render_setup() earlier because of
	you-dont-want-to-know-why. Will change it back once the previews
	have their own render buffers.

	* app/tools/gimpblendtool.c: Replaced the "Type" menu by a preview
	showing the active gradient. Clicking the preview pops up the
	gradient selection. Renamed the "Gradient" menu to "Shape". Removed
	"blend_mode" from the BlendOptions struct because we always use
	"custom" mode now.

+++ app/core/gimpdata.c
@@ -155,6 +155,7 @@ gimp_data_init (GimpData *data)
 {
   data->filename = NULL;
   data->dirty    = FALSE;
+  data->internal = FALSE;
 }
 
 static void
@@ -203,6 +204,12 @@ gimp_data_save (GimpData *data)
 
   g_return_val_if_fail (GIMP_IS_DATA (data), FALSE);
 
+  if (data->internal)
+    {
+      data->dirty = FALSE;
+      return TRUE;
+    }
+
   if (! data->filename)
     {
       g_warning ("%s(): can't save data with NULL filename",
@@ -241,6 +248,9 @@ gimp_data_delete_from_disk (GimpData *data)
   g_return_val_if_fail (GIMP_IS_DATA (data), FALSE);
   g_return_val_if_fail (data->filename != NULL, FALSE);
 
+  if (data->internal)
+    return TRUE;
+
   if (unlink (data->filename) == -1)
     {
       g_message ("%s(): could not unlink() %s: %s",
@@ -269,6 +279,9 @@ gimp_data_set_filename (GimpData    *data,
 			const gchar *filename)
 {
   g_return_if_fail (GIMP_IS_DATA (data));
+
+  if (data->internal)
+    return;
 
   g_free (data->filename);
